Add osd-shared, create osd on mounted devices
Addes logic to create and activate partitions on devices that have
mounted partitions. No changes are made to hooks, the utils library is
updated to handle partitions where it previously bailed.
* Build ceph-disk flags in function to re-use current logic in
osdize_dev
* Shared osd's implement all options/logic of the osdize_dev
 * Checks for in use
 * Handles the format osd option
 * Works with dmcrypt
 * Will not format/zap non-ceph partitions

This includes two drive-by fixes for bugs which it didn't make since to
break out. Both changes are in code which was altered in this
patch and not closing them would require implementing the bug to avoid
closing it.
* Partitions are wiped before zapping
* Zap is done via charmhelpers instead of using the ceph-disk option

Closes-Bug: #1743278
Closes-Bug: #1743280
Closes-Bug: #1743282

Change-Id: I2450a427520b6713abc457374421c94bb7f16078